backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pymongo>=4.8.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.99
pydantic>=2.8.0
